+++ lib/vcr/library_hooks/typhoeus.rb
@@ -2,7 +2,7 @@ require 'vcr/util/version_checker'
 require 'vcr/request_handler'
 require 'typhoeus'
 
-VCR::VersionChecker.new('Typhoeus', Typhoeus::VERSION, '0.2.1', '0.2').check_version!
+VCR::VersionChecker.new('Typhoeus', Typhoeus::VERSION, '0.3.2', '0.3').check_version!
 
 module VCR
   class LibraryHooks
@@ -35,27 +35,17 @@ module VCR
 
       private
 
-        def on_stubbed_request
-          hydra_mock
-        end
-
         def vcr_request
           @vcr_request ||= vcr_request_from(request)
         end
 
-        def typhoeus_response
-          @typhoeus_response ||= ::Typhoeus::Response.new \
+        def on_stubbed_request
+          ::Typhoeus::Response.new \
             :http_version   => stubbed_response.http_version,
             :code           => stubbed_response.status.code,
             :status_message => stubbed_response.status.message,
             :headers_hash   => stubbed_response_headers,
             :body           => stubbed_response.body
-        end
-
-        def hydra_mock
-          @hydra_mock ||= ::Typhoeus::HydraMock.new(/.*/, :any).tap do |m|
-            m.and_return(typhoeus_response)
-          end
         end
 
         def stubbed_response_headers
@@ -75,16 +65,10 @@ module VCR
         end
       end
 
+      ::Typhoeus::Hydra.register_stub_finder do |request|
+        VCR::LibraryHooks::Typhoeus::RequestHandler.new(request).handle
+      end
     end
-  end
-end
-
-# TODO: add Typhoeus::Hydra.register_stub_finder API to Typhoeus
-#       so we can use that instead of monkey-patching it.
-Typhoeus::Hydra::Stubbing::SharedMethods.class_eval do
-  undef find_stub_from_request
-  def find_stub_from_request(request)
-    VCR::LibraryHooks::Typhoeus::RequestHandler.new(request).handle
   end
 end