+++ spec/vcr/library_hooks/typhoeus_spec.rb
@@ -4,9 +4,9 @@ describe "Typhoeus hook", :with_monkey_patches => :typhoeus do
   it_behaves_like 'a hook into an HTTP library', 'typhoeus'
 
   it_performs('version checking', 'Typhoeus',
-    :valid    => %w[ 0.2.1 0.2.99 ],
-    :too_low  => %w[ 0.1.0 0.1.31 0.2.0 ],
-    :too_high => %w[ 0.3.0 1.0.0 ]
+    :valid    => %w[ 0.3.2 0.3.10 ],
+    :too_low  => %w[ 0.2.0 0.2.31 0.3.1 ],
+    :too_high => %w[ 0.4.0 1.0.0 ]
   ) do
     before(:each) { @orig_version = Typhoeus::VERSION }
     after(:each)  { Typhoeus::VERSION = @orig_version }