+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## 2.0.0 Beta 1 (October 8, 2011)
 
+* Update to (and require) Typhoeus 0.3.2.
+
 [Full Changelog](http://github.com/myronmarston/vcr/compare/v1.11.3...v2.0.0.beta1)
 
 ### Changed